`tmpnet`: Write config enabling metrics collection by prometheus
Write prometheus configuration to the ~/.tmpnet/prometheus_sd path for
each node on startup and remove it on shutdown. This enables scraping
of all nodes in a network no matter when they are started.

Ref: https://prometheus.io/docs/guides/file-sd/

Locally tested. Local installation of the monitoring stack is a bit
involved (https://docs.avax.network/nodes/maintain/setting-up-node-monitoring). Getting
this running in CI is the goal but the required persistent prometheus
server has yet to be deployed.